Convert fields to properties in Quest classes

+++ Assets/Scripts/QuestUtils/BasicQuest.cs
@@ -2,6 +2,6 @@ using Microsoft.Geospatial;
 using UnityEngine;
 
 public class BasicQuest : Quest {
-    public BasicQuest(string label, Texture2D referenceImage, float[] featureVector) 
+    public BasicQuest(string label, Texture2D referenceImage, double[] featureVector) 
         : base(label, referenceImage, featureVector) {}
 }
